examples/11_monty_agent_example/pyproject.toml
[project]
name = "monty-agent-example"
version = "0.0.0"
requires-python = ">=3.13"

dependencies = [
  "python-dotenv>=1.0",
  "vision-agents-plugins-gemini",
  "vision-agents-plugins-getstream",
  "vision-agents",
  "pydantic-monty>=0.0.8",
  "aiohttp>=3.9",
]

[tool.uv.sources]
"vision-agents-plugins-gemini" = {path = "../../plugins/gemini", editable=true}
"vision-agents-plugins-getstream" = {path = "../../plugins/getstream", editable=true}
"vision-agents" = {path = "../../agents-core", editable=true}
